Rename back to Identity:: namespace (LG-4231)
- Opposite of #2, now that we have updated the IDP
