+++ docs/contract.md
@@ -2,12 +2,12 @@
 
 ## `/etc/login.gov`
 
-The `/etc/login.gov` directory will exist on deployed instanes of login.gov apps. It contains individual files with useful data, and can be accessed through `LoginGov::Hostdata` class methods
+The `/etc/login.gov` directory will exist on deployed instanes of login.gov apps. It contains individual files with useful data, and can be accessed through `Identity::Hostdata` class methods
 
 | File | API | Example |
 | ---- | --- | ------- |
-| `/etc/login.gov/info/env` | `LoginGov::Hostdata.env` | `"int"` |
-| `/etc/login.gov/info/domain` | `LoginGov::Hostdata.domain` | `"login.gov"` |
+| `/etc/login.gov/info/env` | `Identity::Hostdata.env` | `"int"` |
+| `/etc/login.gov/info/domain` | `Identity::Hostdata.domain` | `"login.gov"` |
 
 ## EC2 instance metadata
 
@@ -16,11 +16,11 @@ We use [instance metadata][instance-metadata] (an HTTP request from an EC2 box) 
 - region
 - account ID
 
-The [LoginGov::Hostdata::EC2](../lib/login_gov/hostdata/ec2.rb) class helps load and read this data:
+The [Identity::Hostdata::EC2](../lib/identity/hostdata/ec2.rb) class helps load and read this data:
 
 ```ruby
-ec2_data = LoginGov::Hostdata::EC2.load
-# => #<LoginGov::Hostdata::EC2:0x00007fc49dd30f40 ...>
+ec2_data = Identity::Hostdata::EC2.load
+# => #<Identity::Hostdata::EC2:0x00007fc49dd30f40 ...>
 ec2_data.region
 # => "us-west-1"
 ec2_data.account_id
@@ -39,10 +39,10 @@ The bucket names can correspond to account IDs and regions:
 "login-gov.app-secrets.#{ec2.account_id}-#{ec2.region}"
 ```
 
-The [LoginGov::Hostdata::S3](../lib/login_gov/hostdata/s3.rb) class helps manage access to data in these buckets, and the `LoginGov::Hostdata.s3` convenience method helps build S3 instances and download remote config files locally:
+The [Identity::Hostdata::S3](../lib/identity/hostdata/s3.rb) class helps manage access to data in these buckets, and the `Identity::Hostdata.s3` convenience method helps build S3 instances and download remote config files locally:
 
 ```ruby
-s3 = LoginGov::Hostdata.s3
+s3 = Identity::Hostdata.s3
 # downloads the `foobar.yml` file for the current environment to local path `config/foobar.yml`
 s3.download_configs(
   '%{env}/idp/v1/foobar.yml' => 'config/foobar.yml'
