+++ lib/identity/hostdata/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
-module LoginGov
+module Identity
   module Hostdata
     VERSION = '0.4.3'
   end
